feat(parity): Particle, terrain, and xfer save/load improvements
Particle System:
- Fixed drift velocity and wind motion handling
- Fixed particle rotation towards emitter
- Fixed alpha/color keyframe timing
- Added priority-based LOD culling support

Terrain System:
- Added Bridge::is_cell_on_side() for pathfinding
- Added Bridge::is_cell_entry_point() for bridge transitions
- Added TerrainLogic::get_extent() for map bounds
- Added add_landmark_bridge_to_logic() and delete_bridge()

Save/Load Xfer:
- Fixed XferVersion type from u32 to u8 (binary compatibility)
- Fixed string serialization to use u8 length prefix (255 char limit)
- Added xfer_drawable_id() method to Xfer trait

+++ GeneralsRust/Code/GameEngine/Common/src/common/rts/production_prerequisite.rs
@@ -604,7 +604,8 @@ impl Snapshotable for ProductionPrerequisite {
     }
 
     fn xfer(&mut self, xfer: &mut dyn Xfer) -> Result<(), String> {
-        const CURRENT_VERSION: u32 = 1;
+        // C++ uses UnsignedByte (u8) for version - matches C++ parity
+        const CURRENT_VERSION: u8 = 1;
         let mut version = CURRENT_VERSION;
         xfer.xfer_version(&mut version, CURRENT_VERSION)
             .map_err(|e| e.to_string())?;
